+++ Examples/Io/Root/include/ActsExamples/Io/Root/RootVertexWriter.hpp
@@ -83,10 +83,10 @@ class RootVertexWriter final : public WriterT<SimVertexContainer> {
   /// Production process type, i.e. what generated the vertex.
   std::vector<std::uint32_t> m_process;
   /// Production position components in mm.
-  std::vector<double> m_vx;
-  std::vector<double> m_vy;
-  std::vector<double> m_vz;
-  std::vector<double> m_vt;
+  std::vector<float> m_vx;
+  std::vector<float> m_vy;
+  std::vector<float> m_vz;
+  std::vector<float> m_vt;
   /// Outgoing particles from the vertex.
   std::vector<std::vector<double>> m_outgoingParticles;
   // Decoded vertex identifier; see Barcode definition for details.
